Tail length calculation added when, both, making the initial repository and when adding a new dog to the existing repository

+++ src/DogClass.java
@@ -20,7 +20,7 @@ public class DogClass {
 	}
 	
     //setter for tail length
-	public void calculateTailLength(String Breed, int Age, int Weight){
+	public void calculateTailLength(String Breed, int Age, double Weight){
 		double length;
 		if( Breed.equals("tax")) {
 			length = 3.7;
